# restapi/src/main/webapp/openapi/openapi.yaml
openapi: 3.0.1
info:
  title: Generator for ProvysObject based classes
  description: Contains methods for geenrating classes, needed to implement access
    to objects from Provys database, described as entities in its metadata catalogue
  version: "1.0"
servers:
- url: /provysobjectgen/api
  variables: {}
paths:
  /entity/{entityNm}/value:
    get:
      summary: Generate Value
      description: Generate value class - object holding all object data
      operationId: getValue
      parameters:
      - name: entityNm
        in: path
        required: true
        schema:
          type: string
      - name: friendEntities
        in: query
        schema:
          type: string
      responses:
        default:
          description: Class source
          content:
            text/plain:
              schema:
                type: string
  /entity/{entityNm}/proxy:
    get:
      summary: Generate Proxy
      description: Generate proxy - simple envelope on generated proxy file. Used
        only for initialisation, later not regenerated
      operationId: getProxy
      parameters:
      - name: entityNm
        in: path
        required: true
        schema:
          type: string
      - name: friendEntities
        in: query
        schema:
          type: string
      responses:
        default:
          description: Class source
          content:
            text/plain:
              schema:
                type: string
  /entity/{entityNm}/geninterface:
    get:
      summary: Generate GenInterface
      description: Generate generated ancestor of interface for accessing objects
        of given type, containing attribute getters
      operationId: getGenInterface
      parameters:
      - name: entityNm
        in: path
        required: true
        schema:
          type: string
      - name: friendEntities
        in: query
        schema:
          type: string
      responses:
        default:
          description: Class source
          content:
            text/plain:
              schema:
                type: string
  /entity/{entityNm}/loaderbase:
    get:
      summary: Generate Loader Base class
      description: Generate loader base class - common ancestor for loaders for given
        entity.
      operationId: getLoaderBase
      parameters:
      - name: entityNm
        in: path
        required: true
        schema:
          type: string
      - name: friendEntities
        in: query
        schema:
          type: string
      responses:
        default:
          description: Class source
          content:
            text/plain:
              schema:
                type: string
  /entity/{entityNm}/dbloadrunner:
    get:
      summary: Generate DbLoadRunner class
      description: Generate database load runner class - class that actually fetches
        data from database and converts them to value objects.
      operationId: getDbLoadRunner
      parameters:
      - name: entityNm
        in: path
        required: true
        schema:
          type: string
      - name: friendEntities
        in: query
        schema:
          type: string
      responses:
        default:
          description: Class source
          content:
            text/plain:
              schema:
                type: string
  /entity/{entityNm}/loader:
    get:
      summary: Generate Loader Interface
      description: Generate loader interface - interface defining requirements for
        loader for given entity. Only used to initialize file if one doesn't exist,
        is not refreshed later
      operationId: getLoaderInterface
      parameters:
      - name: entityNm
        in: path
        required: true
        schema:
          type: string
      - name: friendEntities
        in: query
        schema:
          type: string
      responses:
        default:
          description: Class source
          content:
            text/plain:
              schema:
                type: string
  /entity/{entityNm}/xmlvalueadapter:
    get:
      summary: Generate JAXB Value Adapter
      description: Generate JAXB value adapter - translates value to value builder
        for serialisation
      operationId: getXmlValueAdapter
      parameters:
      - name: entityNm
        in: path
        required: true
        schema:
          type: string
      - name: friendEntities
        in: query
        schema:
          type: string
      responses:
        default:
          description: Class source
          content:
            text/plain:
              schema:
                type: string
  /entity/{entityNm}/dbloader:
    get:
      summary: Generate DbLoader class
      description: Generate database loader class - at the moment, does not generate
        load by methods.
      operationId: getDbLoader
      parameters:
      - name: entityNm
        in: path
        required: true
        schema:
          type: string
      - name: friendEntities
        in: query
        schema:
          type: string
      responses:
        default:
          description: Class source
          content:
            text/plain:
              schema:
                type: string
  /entity/{entityNm}/interface:
    get:
      summary: Generate Interface
      description: Generate interface, representing instance of object of given type
      operationId: getInterface
      parameters:
      - name: entityNm
        in: path
        required: true
        schema:
          type: string
      - name: friendEntities
        in: query
        schema:
          type: string
      responses:
        default:
          description: Class source
          content:
            text/plain:
              schema:
                type: string
  /entity/{entityNm}/valuebuilder:
    get:
      summary: Generate Value Builder
      description: Generate value builder class - object allowing creation or modification
        of object value
      operationId: getValueBuilder
      parameters:
      - name: entityNm
        in: path
        required: true
        schema:
          type: string
      - name: friendEntities
        in: query
        schema:
          type: string
      responses:
        default:
          description: Class source
          content:
            text/plain:
              schema:
                type: string
  /entity/{entityNm}/genproxy:
    get:
      summary: Generate GenProxy
      description: Generate generated proxy ancestor, implementing interface of given
        object in scope defined by generated interface
      operationId: getGenProxy
      parameters:
      - name: entityNm
        in: path
        required: true
        schema:
          type: string
      - name: friendEntities
        in: query
        schema:
          type: string
      responses:
        default:
          description: Class source
          content:
            text/plain:
              schema:
                type: string
  /entity/{entityNm}/xmlproxyadapter:
    get:
      summary: Generate JAXB Proxy Adapter
      description: Generate JAXB proxy adapter - translates proxy to value object
        for serialisation
      operationId: getXmlProxyAdapter
      parameters:
      - name: entityNm
        in: path
        required: true
        schema:
          type: string
      - name: friendEntities
        in: query
        schema:
          type: string
      responses:
        default:
          description: Class source
          content:
            text/plain:
              schema:
                type: string
  /entity/{entityNm}/jsonbproxyadapter:
    get:
      summary: Generate JSON-B Proxy Adapter
      description: Generate JSON-B proxy adapter - translates proxy to value object
        for serialisation
      operationId: getJsonbProxyAdapter
      parameters:
      - name: entityNm
        in: path
        required: true
        schema:
          type: string
      - name: friendEntities
        in: query
        schema:
          type: string
      responses:
        default:
          description: Class source
          content:
            text/plain:
              schema:
                type: string
  /entity/{entityNm}/jsonbvalueadapter:
    get:
      summary: Generate JSON-B Value Adapter
      description: Generate JSON-B value adapter - translates value to value builder
        for serialisation
      operationId: getJsonbValueAdapter
      parameters:
      - name: entityNm
        in: path
        required: true
        schema:
          type: string
      - name: friendEntities
        in: query
        schema:
          type: string
      responses:
        default:
          description: Class source
          content:
            text/plain:
              schema:
                type: string